+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 ### Added
 ### Fixed
-- all fundamental callbacks now raise an error if not implemented
+- getTermsQuadratic() now correctly returns all linear terms
 ### Changed
 ### Removed